refactor(api): move set package
