+++ api/internal/media/entity/regulation.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	set "github.com/and-period/furumaru/api/pkg/set/v2"
+	"github.com/and-period/furumaru/api/pkg/set"
 	"github.com/and-period/furumaru/api/pkg/uuid"
 )
