+++ api/internal/store/entity/live.go
@@ -3,7 +3,7 @@ package entity
 import (
 	"time"
 
-	set "github.com/and-period/furumaru/api/pkg/set/v2"
+	"github.com/and-period/furumaru/api/pkg/set"
 	"github.com/and-period/furumaru/api/pkg/uuid"
 	"gorm.io/gorm"
 )
