+++ api/internal/store/entity/live_product.go
@@ -3,7 +3,7 @@ package entity
 import (
 	"time"
 
-	set "github.com/and-period/furumaru/api/pkg/set/v2"
+	"github.com/and-period/furumaru/api/pkg/set"
 )
 
 // ライブ配信関連商品情報
