+++ api/pkg/set/set.go
@@ -1,140 +1,107 @@
 package set
 
-import "sort"
-
 // Set - 重複排除処理用の構造体
-type Set struct {
-	values map[interface{}]struct{}
+type Set[T comparable] struct {
+	values map[T]struct{}
 }
 
 // New - 構造体の生成(cap指定)
-func New(cap int) *Set {
-	return &Set{
-		values: make(map[interface{}]struct{}, cap),
+func New[T comparable](values ...T) *Set[T] {
+	return NewEmpty[T](len(values)).Add(values...)
+}
+
+func NewEmpty[T comparable](cap int) *Set[T] {
+	return &Set[T]{
+		values: make(map[T]struct{}, cap),
 	}
 }
 
-// Set - 構造体の中身を初期化
-func (s *Set) Reset(cap int) {
-	s.values = make(map[interface{}]struct{}, cap)
+// Uniq - 渡された値の重複を排除して返す
+func Uniq[T comparable](values ...T) []T {
+	return New(values...).Slice()
+}
+
+// UniqBy - 渡された操作を実行して重複を排除した値を返す
+func UniqBy[K comparable, V any](values []V, iteratee func(V) K) []K {
+	set := NewEmpty[K](len(values))
+	for i := range values {
+		key := iteratee(values[i])
+		set.Add(key)
+	}
+	return set.Slice()
+}
+
+// UniqWithErr - 渡された操作を実行して重複を排除した値を返す
+func UniqWithErr[K comparable, V any](values []V, iteratee func(V) (K, error)) ([]K, error) {
+	set := NewEmpty[K](len(values))
+	for i := range values {
+		key, err := iteratee(values[i])
+		if err != nil {
+			return nil, err
+		}
+		set.Add(key)
+	}
+	return set.Slice(), nil
 }
 
 // Len - 長さを取得
-func (s *Set) Len() int {
+func (s *Set[T]) Len() int {
 	return len(s.values)
 }
 
-// Contains - 指定された値がすべて含まれているかの判定
-func (s *Set) Contains(values ...interface{}) bool {
+// Reset - 構造体の中身を初期化
+func (s *Set[T]) Reset(cap int) *Set[T] {
+	s.values = make(map[T]struct{}, cap)
+	return s
+}
+
+// Contains - 指定された値がすべて含まれているかの検証
+func (s *Set[T]) Contains(values ...T) bool {
 	for i := range values {
-		if _, ok := s.values[values[i]]; !ok {
-			return false
+		if _, ok := s.values[values[i]]; ok {
+			continue
 		}
+		return false
 	}
 	return true
 }
 
+// Add - 指定された値を代入
+func (s *Set[T]) Add(values ...T) *Set[T] {
+	for _, v := range values {
+		s.values[v] = struct{}{}
+	}
+	return s
+}
+
 // FindOrAdd - 指定された値のいずれかが存在するかの判定 && 存在しないものは代入
-func (s *Set) FindOrAdd(values ...interface{}) bool {
+func (s *Set[T]) FindOrAdd(values ...T) (*Set[T], bool) {
 	var isExists bool
 	for i := range values {
-		if _, ok := s.values[values[i]]; ok {
+		if s.Contains(values[i]) {
 			isExists = true
 			continue
 		}
 		s.Add(values[i])
 	}
-	return isExists
-}
-
-// Add - 指定された値を代入
-func (s *Set) Add(values ...interface{}) {
-	for _, v := range values {
-		s.values[v] = struct{}{}
-	}
-}
-
-// AddStrings - 指定された文字列を代入
-func (s *Set) AddStrings(values ...string) {
-	for _, v := range values {
-		s.Add(v)
-	}
-}
-
-// AddInt32s - 指定された数字(int32)を代入
-func (s *Set) AddInt32s(values ...int32) {
-	for _, v := range values {
-		s.Add(v)
-	}
-}
-
-// AddInt64s - 指定された数字(int64)を代入
-func (s *Set) AddInt64s(values ...int64) {
-	for _, v := range values {
-		s.Add(v)
-	}
+	return s, isExists
 }
 
 // Remove - 指定された値を削除
-func (s *Set) Remove(v interface{}) {
+func (s *Set[T]) Remove(v T) *Set[T] {
 	delete(s.values, v)
+	return s
 }
 
-// Do - 構造体内の内のデータに対し、指定された処理を実行
-func (s *Set) Do(f func(interface{})) {
+// Slice - Sliceとして返す
+func (s *Set[T]) Slice() []T {
+	var zero T
+	res := make([]T, 0, s.Len())
 	for v := range s.values {
-		f(v)
+		if v == zero {
+			continue
+		}
+		res = append(res, v)
 	}
-}
-
-// Strings - 文字列型の配列として返す
-func (s *Set) Strings() []string {
-	res := make([]string, 0, s.Len())
-	for v := range s.values {
-		res = append(res, v.(string))
-	}
-	return res
-}
-
-// SortStrings - 文字列型の配列 (昇順でソートされた状態) として返す
-func (s *Set) SortStrings() []string {
-	res := s.Strings()
-	sort.Strings(res)
-	return res
-}
-
-// Int32s - int32型の配列として返す
-func (s *Set) Int32s() []int32 {
-	res := make([]int32, 0, s.Len())
-	for v := range s.values {
-		res = append(res, v.(int32))
-	}
-	return res
-}
-
-// SortInt32s - int32型の配列 (昇順でソートされた状態) として返す
-func (s *Set) SortInt32s() []int32 {
-	res := s.Int32s()
-	sort.Slice(res, func(i, j int) bool {
-		return res[i] < res[j]
-	})
-	return res
-}
-
-// Int64s - int64型の配列として返す
-func (s *Set) Int64s() []int64 {
-	res := make([]int64, 0, s.Len())
-	for v := range s.values {
-		res = append(res, v.(int64))
-	}
-	return res
-}
-
-// SortInt64s - int64型の配列 (昇順でソートされた状態) として返す
-func (s *Set) SortInt64s() []int64 {
-	res := s.Int64s()
-	sort.Slice(res, func(i, j int) bool {
-		return res[i] < res[j]
-	})
 	return res
 }
